medical-rag/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
sqlmodel==0.0.14
aiofiles==23.2.1
httpx==0.26.0
python-multipart==0.0.6
beautifulsoup4==4.12.3
lxml==5.1.0
openai==1.12.0
faiss-cpu==1.7.4
chromadb==0.4.22
sentence-transformers==2.3.1
tiktoken==0.6.0
python-dotenv==1.0.0
pytest==8.0.0
pytest-asyncio==0.23.4
